remove duplicate transformations + add all_for_ingestions
