+++ apps/andi/test/integration/andi/input_schemas/ingestions/transformations_test.exs
@@ -13,6 +13,7 @@ defmodule Andi.InputSchemas.Ingestions.TransformationsTest do
   describe "get/1" do
     test "returns a saved transformation by id" do
       transformation_id = UUID.uuid4()
+
       changes = %{
         id: transformation_id,
         name: "sample transformation",
@@ -24,17 +25,18 @@ defmodule Andi.InputSchemas.Ingestions.TransformationsTest do
         }
       }
 
-      changes |> Transformation.changeset()
+      changes
+      |> Transformation.changeset()
       |> Transformations.save()
 
       assert %{
-                id: ^transformation_id,
-                type: "concatenation",
-                parameters: %{
-                  "sourceFields" => ["other", "name"],
-                  "separator" => ".",
-                  "targetField" => "name"
-                }
+               id: ^transformation_id,
+               type: "concatenation",
+               parameters: %{
+                 "sourceFields" => ["other", "name"],
+                 "separator" => ".",
+                 "targetField" => "name"
+               }
              } = Transformations.get(transformation_id)
     end
   end
@@ -64,6 +66,84 @@ defmodule Andi.InputSchemas.Ingestions.TransformationsTest do
       eventually(fn ->
         assert nil == Transformations.get(id)
       end)
+    end
+  end
+
+  # todo: get all
+  # todo: update
+
+  describe "all_for_ingestion/1" do
+    test "given existing transformations, it returns them" do
+      dataset_one = TDG.create_dataset(%{})
+      dataset_two = TDG.create_dataset(%{})
+
+      assert {:ok, _} = Datasets.update(dataset_one)
+      assert {:ok, _} = Datasets.update(dataset_two)
+
+      ingestion_one =
+        TDG.create_ingestion(%{
+          targetDataset: dataset_one.id,
+          transformations: [
+            %{
+              type: "concatenation",
+              parameters: %{
+                "sourceFields" => ["other", "name"],
+                "separator" => ".",
+                "targetField" => "name"
+              }
+            },
+            %{
+              type: "regex_extract",
+              parameters: %{
+                "sourceField" => "name",
+                "targetField" => "firstName",
+                "regex" => "^(\\w+)"
+              }
+            }
+          ]
+        })
+
+      ingestion_two =
+        TDG.create_ingestion(%{
+          targetDataset: dataset_two.id,
+          transformations: [
+            %{
+              type: "regex_extract",
+              parameters: %{
+                "sourceField" => "name",
+                "targetField" => "firstName",
+                "regex" => "^(\\w+)"
+              }
+            }
+          ]
+        })
+
+      assert {:ok, _} = Ingestions.update(ingestion_one)
+      assert {:ok, _} = Ingestions.update(ingestion_two)
+
+      ingestion_one_id = ingestion_one.id
+      ingestion_two_id = ingestion_two.id
+
+      assert [
+               %{
+                 type: "concatenation",
+                 parameters: %{
+                   "sourceFields" => ["other", "name"],
+                   "separator" => ".",
+                   "targetField" => "name"
+                 },
+                 ingestion_id: ^ingestion_one_id
+               },
+               %{
+                 type: "regex_extract",
+                 parameters: %{
+                   "sourceField" => "name",
+                   "targetField" => "firstName",
+                   "regex" => "^(\\w+)"
+                 },
+                 ingestion_id: ^ingestion_one_id
+               }
+             ] = Transformations.all_for_ingestion(ingestion_one_id)
     end
   end
 end
